move all index pages up a level and name them after their folder
