+++ pages/index.tsx
@@ -16,7 +16,6 @@ import imgHomeConsoleBlogLogo from '../public/images/icons/blog/ios/AppIcon@3x.p
 import imgHomeDoFLogo from '../public/images/icons/depth-of-field/ios/AppIcon@3x.png'
 import imgWithinViewportFLogo from '../public/images/icons/within-viewport/ios/AppIcon@3x.png'
 import imgTravelParis from '../public/images/travel-paris.jpg'
-import IntroFormatted from '../src/components/IntroFormatted'
 import Layout from '../src/components/Layout'
 
 export default function Home() {
@@ -26,12 +25,10 @@ export default function Home() {
                 <link rel="me" href="https://mastodon.online/@craigpatik" key="mastodon" />
             </Head>
             <div role="main">
-                <IntroFormatted>
-                    <p>
-                        I’m <Link href="/about">Craig Patik</Link>, a traveler, web developer, and Arsenal fan based in
-                        Munich, Germany.
-                    </p>
-                </IntroFormatted>
+                <p>
+                    I’m <Link href="/about">Craig Patik</Link>, a traveler, web developer, and Arsenal fan based in
+                    Munich, Germany.
+                </p>
 
                 <section>
                     <h2>Travel</h2>
