+++ styles/scss/_dark-mode.scss
@@ -1,3 +1,8 @@
+body {
+    background-color: white;
+    color: $dark-charcoal;
+}
+
 @media (prefers-color-scheme: dark) {
     body {
         background-color: $dark-charcoal;
