+++ styles/scss/_foundation-overrides.scss
@@ -92,19 +92,6 @@ $portrait: 'only screen and (orientation: portrait)';
 // $total-columns: 12;
 
 //
-// Block Grid Variables
-//
-
-// We use this to control the maximum number of block grid elements per row
-
-// $block-grid-elements: 12;
-// $block-grid-default-spacing: em-calc(20);
-
-// Enables media queries for block-grid classes. Set to false if writing semantic HTML.
-
-// $block-grid-media-queries: true;
-
-//
 // Typography Variables
 //
